Moved type_traits tests, added tests for enable_if, slight tweak to type_traits to avoid user code clashes
